+++ doc/changelog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## 3.2.2 - 2017-01-06 ##
 ### Fixed
 * Defer now use fit on change
+* Wrong class added on slider when unique
 
 ## 3.2.1 - 2017-01-02 ##
 ### Fixed